Starting to move DOFBases and RefFEs
