+++ src/REFACTORED/RefFEs/files.jl
@@ -4,3 +4,9 @@ include("PolynomialBases.jl")
 
 include("Polytopes.jl")
 @reexport using Gridap.Polytopes
+
+#include("DOFBases.jl")
+#@reexport using Gridap.DOFBases
+#
+#include("RefFEs.jl")
+#@reexport using Gridap.RefFEs
