+++ test/REFACTORED/RefFEsTests/DOFBasesTests.jl
@@ -1,0 +1,6 @@
+module DOFBasesTests
+
+using Test
+using Gridap
+
+end # module
