+++ test/REFACTORED/RefFEsTests/runtests.jl
@@ -6,4 +6,6 @@ using Test
 
 @testset "Polytopes" begin include("PolytopesTests.jl") end
 
+#@testset "DOFBases" begin include("DOFBasesTests.jl") end
+
 end # module
